// packages/agent-runtime/src/codex/generated/codex-app-server/schema/v2/PermissionsRequestApprovalResponse.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GrantedPermissionProfile } from "./GrantedPermissionProfile.js";
import type { PermissionGrantScope } from "./PermissionGrantScope.js";

export type PermissionsRequestApprovalResponse = { permissions: GrantedPermissionProfile, scope: PermissionGrantScope, };